拆分linkedList
