+++ .gitignore
@@ -1,4 +1,6 @@
+
 build
+bin
 .DS_Store
 *.o
 *.a
